+++ lifx/const.go
@@ -1,12 +1,14 @@
 package lifx
 
 const (
-	//ActionTurnOnLight is the UUID to turn on the light
-	ActionTurnOnLight = "dev.sean.lifx.on"
-	//ActionTurnOffLight is the UUID to turn on the light
-	ActionTurnOffLight = "dev.sean.lifx.off"
-	//ActionSetColor is the UUID to turn on the light
+	//ActionTurnOnDevice is the UUID to turn on the device
+	ActionTurnOnDevice = "dev.sean.lifx.on"
+	//ActionTurnOffDevice is the UUID to turn on the device
+	ActionTurnOffDevice = "dev.sean.lifx.off"
+	//ActionSetColor is the UUID to set the color on the device
 	ActionSetColor = "dev.sean.lifx.color"
-	//ActionSetBrightness is the UUID to turn on the light
+	//ActionSetBrightness is the UUID to set the brightness on the device
 	ActionSetBrightness = "dev.sean.lifx.brightness"
+	//ActionSetWaveform is the UUID to set a waveform on the device
+	ActionSetWaveform = "dev.sean.lifx.waveform"
 )