+++ lifx/streamdeck.go
@@ -18,39 +18,18 @@ func (c *Client) OnWillAppear(msg streamdeck.WillAppearMsg) {
 		//this is brand new or unconfigured
 		return
 	}
-	switch msg.Action {
-	case ActionSetColor:
-		if val, ok := msg.Payload.Settings["version"]; ok {
-			// Version set so work on migrations
-			switch val {
-
-			}
-		} else {
-			// Version was not set, so migrate to latest version
-			settings := migrateColorSettingsToV1(msg.Payload.Settings)
-			settingsMsg := streamdeck.SetSettingsMsg{
-				Context: msg.Context,
-				Event:   streamdeck.SetSettingsEvent,
-				Payload: settings,
-			}
-			c.sdClient.SendMessage(settingsMsg)
-		}
-	case ActionSetBrightness:
-		if val, ok := msg.Payload.Settings["version"]; ok {
-			// version set so work on migrations if needed
-			switch val {
-
-			}
-		} else {
-			settings := migrateBrightnessSettingsToV1(msg.Payload.Settings)
-			settingsMsg := streamdeck.SetSettingsMsg{
-				Context: msg.Context,
-				Event:   streamdeck.SetSettingsEvent,
-				Payload: settings,
-			}
-			c.sdClient.SendMessage(settingsMsg)
-		}
+	// Migrate settings if needed
+	settings, err := c.Migrate(msg.Payload.Settings, msg.Action)
+	if err != nil {
+		c.sdClient.Log(err.Error())
+		panic(err)
 	}
+	settingsMsg := streamdeck.SetSettingsMsg{
+		Context: msg.Context,
+		Event:   streamdeck.SetSettingsEvent,
+		Payload: settings,
+	}
+	c.sdClient.SendMessage(settingsMsg)
 }
 
 // OnSendToPlugin is called when a message is sent to the plugin (generally from the property inspector)
@@ -66,14 +45,16 @@ func (c *Client) OnSendToPlugin(msg streamdeck.SendToPluginMsg) {
 // OnKeyUp is called when the Stream Deck key is unpressed
 func (c *Client) OnKeyUp(msg streamdeck.KeyUpMsg) {
 	switch msg.Action {
-	case ActionTurnOnLight:
+	case ActionTurnOnDevice:
 		c.turnLights(msg.Context, msg.Payload.Settings, true)
-	case ActionTurnOffLight:
+	case ActionTurnOffDevice:
 		c.turnLights(msg.Context, msg.Payload.Settings, false)
 	case ActionSetColor:
 		c.setColor(msg.Context, msg.Payload.Settings)
 	case ActionSetBrightness:
 		c.setBrightness(msg.Context, msg.Payload.Settings)
+	case ActionSetWaveform:
+		c.setWaveform(msg.Context, msg.Payload.Settings)
 	default:
 		c.SendWarnMessage(msg.Context)
 		c.sdClient.Log(fmt.Sprintf("Unknown action received %s", msg.Action))